default values sample in one template and separate html files for documents and templates

+++ src/repositories/documents/in-memory/documents.ts
@@ -1,8 +1,8 @@
 import * as handlebars from 'handlebars';
-import * as _ from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
 import * as fs from 'fs';
 import { Express } from 'express';
+import axios from 'axios';
 import { AttachmentJSON } from '@sendgrid/helpers/classes/attachment';
 import { DocumentsInterface } from '../../../models/interfaces/documents';
 import { Template } from '../../../models/types/template';
@@ -12,26 +12,34 @@ import sendEmail from '../../../utils/email';
 
 export default class InMemoryDocumentsRepo implements DocumentsInterface {
   // eslint-disable-next-line no-unused-vars
-  async createDocument(data : any, template : Template, recommendationId : string, docName : string, userId : string) : Promise<{ document : Document }> {
-    const templateFields = template.questions.map((value) => value.field);
-    const dataFields = Object.keys(data);
-    if (_.isEqual(templateFields.sort(), dataFields.sort())) {
-      const { html } = template;
-      const compiledHtml = handlebars.compile(html);
-      const document : Document = {
-        id: uuidv4(),
-        template_id: template.templateId,
-        name: docName,
-        createdOn: new Date(),
-        createdBy: userId,
-        html: compiledHtml(data),
-      };
-      fileCheck(`${__dirname}/data`, false);
-      fs.writeFileSync(`${__dirname}/data/${document.id}.json`, JSON.stringify(document));
-      return { document };
+  async createDocument(data : any, required : any, template : Template, recommendationId : string, docName : string, userId : string) : Promise<{ document : Document }> {
+    let requiredResources = {};
+    if (template.resources) {
+      // eslint-disable-next-line no-restricted-syntax
+      for await (const resource of template.resources) {
+        const compiledUrl = handlebars.compile(resource.url);
+        requiredResources[resource.name] = (await axios.get(compiledUrl(required), { headers: { userid: '1238989' } })).data;
+      }
     }
-    const err = { message: 'All fields are required', statusCode: 400 };
-    throw err;
+    requiredResources = {
+      ...requiredResources,
+      ...data,
+    };
+    const { html } = template;
+    const compiledHtml = handlebars.compile(fs.readFileSync(`src/repositories/templates/in-memory/data/html/${html}`).toString().replace(/\n/g, '').replace(/(^"|"$)/g, ''));
+    const document : Document = {
+      id: uuidv4(),
+      template_id: template.templateId,
+      name: docName,
+      createdOn: new Date(),
+      createdBy: userId,
+      html,
+    };
+    fileCheck(`${__dirname}/data/html`, false);
+    fs.writeFileSync(`${__dirname}/data/html/${document.id}.html`, JSON.stringify(compiledHtml(requiredResources)));
+    fileCheck(`${__dirname}/data`, false);
+    fs.writeFileSync(`${__dirname}/data/${document.id}.json`, JSON.stringify(document));
+    return { document };
   }
 
   async getAllDocuments() : Promise<Document[]> {
@@ -60,6 +68,7 @@ export default class InMemoryDocumentsRepo implements DocumentsInterface {
     if (fs.existsSync(`${__dirname}/data/${id}.json`)) {
       const readData = fs.readFileSync(`${__dirname}/data/${id}.json`).toString();
       const data = JSON.parse(readData) as Document;
+      data.html = fs.readFileSync(`${__dirname}/data/html/${id}.html`).toString().replace(/\n/g, '').replace(/(^"|"$)/g, '');
       return data;
     }
     const err = { message: `Document not found for id: ${id}`, statusCode: 404 };
@@ -67,8 +76,9 @@ export default class InMemoryDocumentsRepo implements DocumentsInterface {
   }
 
   async deleteDocument(id : string) : Promise<{ success : boolean }> {
-    if (fs.existsSync(`${__dirname}/data/${id}.json`)) {
+    if (fs.existsSync(`${__dirname}/data/${id}.json`) && fs.existsSync(`${__dirname}/data/html/${id}.html`)) {
       fs.unlinkSync(`${__dirname}/data/${id}.json`);
+      fs.unlinkSync(`${__dirname}/data/html/${id}.html`);
       return { success: true };
     }
     const err = { message: `Document not found for id: ${id}`, statusCode: 404 };
@@ -76,11 +86,8 @@ export default class InMemoryDocumentsRepo implements DocumentsInterface {
   }
 
   async editDocument(id : string, html : string) : Promise<{ success : boolean }> {
-    if (fs.existsSync(`${__dirname}/data/${id}.json`)) {
-      const jsonString = fs.readFileSync(`${__dirname}/data/${id}.json`).toString();
-      const jsonDocument = JSON.parse(jsonString) as Document;
-      jsonDocument.html = html;
-      fs.writeFileSync(`${__dirname}/data/${id}.json`, JSON.stringify(jsonDocument));
+    if (fs.existsSync(`${__dirname}/data/${id}.json`) && fs.existsSync(`${__dirname}/data/html/${id}.html`)) {
+      fs.writeFileSync(`${__dirname}/data/html/${id}.html`, JSON.stringify(html));
       return { success: true };
     }
     const err = { message: `Document not found for id: ${id}`, statusCode: 404 };
